+++ CHANGELOG.md
@@ -11,6 +11,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   report on the value that caused the error (unlike the errors previously
   converted from standard library e.g. core::num::ParseIntError).
 
+- Added: new expression functions:
+
+  * `defined(v)` - predicate to determine if variable 'v' is defined.
+  * `var(v)` - return the (string) value of variable 'v'
+  * `num(x)` - convert the string x into a number
+  
+  The last two of these can throw errors: `Undefined` for `var()` and
+  `ParseNumber` for `num()`.
+
 - Added: `reformat()` library method to reformat an input document without
   transforming it. Exposed through the CLI as `--reformat-only`, and in the
   editor via the Input > Reformat menu option.